+++ kitty/layout/interface.py
@@ -7,13 +7,14 @@ from .grid import Grid
 from .splits import Splits
 from .stack import Stack
 from .tall import Fat, Tall
-from .vertical import Horizontal, Vertical
+from .vertical import Horizontal, Vertical, VerticalFocus
 
 all_layouts: dict[str, type[Layout]] = {
     Stack.name: Stack,
     Tall.name: Tall,
     Fat.name: Fat,
     Vertical.name: Vertical,
+    VerticalFocus.name: VerticalFocus,
     Horizontal.name: Horizontal,
     Grid.name: Grid,
     Splits.name: Splits,